Modularization of components as well as restructuring of application
