+++ styles/colors.js
@@ -1,0 +1,10 @@
+const colors = {
+  darkGrey: "#222",
+  lightGrey: "#e9e9e9",
+  emerald: "#2ecc71",
+  lightEmerald: "#b1dfbb",
+  gold: "#f6b01c",
+  blue: "#3498db"
+};
+
+export default colors;
